moves card window to own module
