+++ gui/src/card_window/card_view_model.rs
