+++ gui/src/card_window/mod.rs
@@ -1,0 +1,2 @@
+pub(super) mod card_window;
+pub(super) mod card_view_model;
