+++ gui/src/card_window/readme.md
@@ -1,0 +1,3 @@
+# Card Window
+
+Inside this module all the things to display and interact with a card are defined.
